Add cinema room statistics and seat validation
Expanded the functionalities of the Cinema Room Manager. This includes the addition of statistics information providing the number of purchased tickets, capacity percentage, current and total income. Moreover, seat validation was implemented to ensure that double booking of a previously purchased seat or entering incorrect seat information is handled gracefully.